Renamed to oauthproxy to better capture intent
While authsome is a great name, this is the wrong project to use it
for. This project has been tidied up and hopefully now is focussed
on providing just what is was intended to.

The downside of the big rename is that this is now untested code;
I've not gone back to make sure this all works.